+++ lib/help_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+// view for FAQs
+class HelpPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Text('How does it work?'),
+        Text('FAQ'),
+      ],
+    );
+  }
+}
